Refactor: Rename `uri` to `imageUri` and `menu` to `optionsMenu`
This commit renames the `uri` parameter to `imageUri` in `ImageProcessor.kt` and its usages throughout the test files for clarity.

Additionally, the `menu` parameter in `MainActivity.kt`'s `onCreateOptionsMenu` method is renamed to `optionsMenu` to better reflect its purpose.

In `FirstFragment.kt`, the `maskMat` parameter in `createMaskOverlay` is renamed to `maskMatrix`.